models made

+++ store/models.py
@@ -1,6 +1,5 @@
 from django.db import models
 from django.contrib.auth.models import User
-from django.db.models.expressions import F
 # Create your models here.
 
 
@@ -16,7 +15,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=200)
-    price = models.DecimalField()
+    price = models.FloatField()
     digital = models.BooleanField(default=False, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
 
@@ -25,15 +24,16 @@ class Product(models.Model):
 
     @property
     def imageURL(self):
-           try:
-                url = self.image.url
-           except:
-                url = ''
-           return url
+        try:
+            url = self.image.url
+        except:
+            url = ''
+        return url
 
 
 class Order(models.Model):
-    customer = models.ForeignKey(Customer, on_delete=models.SET_NULL, null=True,blank=True)
+    customer = models.ForeignKey(
+        Customer, on_delete=models.SET_NULL, null=True, blank=True)
     date_ordered = models.DateField(auto_now_add=True)
     complete = models.BooleanField(default=False)
     transaction_id = models.CharField(max_length=200, null=True)
@@ -41,26 +41,26 @@ class Order(models.Model):
     def __str__(self):
         return str(self.id)
 
-	@property
-	def shipping(self):
-	        shipping = False
-		orderitems = self.orderitem_set.all()
-		for i in orderitems:
-			        if i.product.digital == False:
-				        shipping = True
-		return shipping
+    @property
+    def shipping(self):
+            shipping = False
+            orderitems = self.orderitem_set.all()
+            for i in orderitems:
+                if i.product.digital == False:
+                    shipping = True
+            return shipping
 
     @property
     def get_cart_items(self):
-            orderitems = self.orderitem_set.all()
-            total = sum([item.quantity for item in orderitems])
-            return total             
-    
+        orderitems = self.orderitem_set.all()
+        total = sum([item.quantity for item in orderitems])
+        return total
+
     @property
     def get_cart_total(self):
-           orderitems = self.orderitem_set.all()
-           total = sum([item.get_total for item in orderitems])
-           return total
+        orderitems = self.orderitem_set.all()
+        total = sum([item.get_total for item in orderitems])
+        return total
 
 
 class OrderItem(models.Model):
